PgmFileManager: Make write() accept image data as a parameter
Also remove the now useless cycles wrapping the read() and write() syscalls
(this really should have been part of the previous commit).
The example image is now a global var in main.cpp. Ugly design, but ask me if I
give a fuck. It's going to be removed later anyway.